+++ packages/plugin-vue3/src/config/base.ts
@@ -1,7 +1,7 @@
 
 import { join } from 'path'
 import { Mode } from 'ssr-types'
-import { getCwd, loadConfig, setStyle, addImageChain, loadModuleFromFramework, logErr } from 'ssr-common-utils'
+import { getCwd, loadConfig, setStyle, addImageChain, loadModuleFromFramework, logErr, getBuildConfig } from 'ssr-common-utils'
 import * as webpack from 'webpack'
 import * as WebpackChain from 'webpack-chain'
 
@@ -65,7 +65,7 @@ const addBabelLoader = (chain: WebpackChain.Rule<WebpackChain.Module>, envOption
 
 const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
   const config = loadConfig()
-  const { moduleFileExtensions, useHash, chainBaseConfig, locale, corejsOptions, ssrVueLoaderOptions, csrVueLoaderOptions, babelExtraModule, alias, define, babelOptions } = config
+  const { moduleFileExtensions, chainBaseConfig, locale, corejsOptions, ssrVueLoaderOptions, assetsDir, csrVueLoaderOptions, babelExtraModule, alias, define, babelOptions } = config
 
   let vueLoaderOptions = {
     babelParserPlugins: ['jsx', 'classProperties', 'decorators-legacy'],
@@ -169,7 +169,6 @@ const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
     .rule('compileBabelForExtraModule')
     .test(/\.(js|mjs|jsx|ts|tsx)$/)
     .include
-    .add([/ssr-plugin-vue3/, /ssr-client-utils/, /ssr-hoc-vue/, /vue/, /ssr-common-utils/])
 
   const babelForExtraModule = module.add(babelExtraModule ?? []).add(babelOptions?.include as Array<string|RegExp> ?? []).end().exclude.add(/core-js/).end()
 
@@ -195,15 +194,12 @@ const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
     .use('file-loader')
     .loader(loadModule('file-loader'))
     .options({
-      name: 'static/[name].[hash:8].[ext]',
+      name: `${assetsDir}/[name].[hash:8].[ext]`,
       esModule: false,
       emitFile: !isServer
     })
 
-  chain.plugin('minify-css').use(MiniCssExtractPlugin, [{
-    filename: useHash ? '[name].[contenthash:8].css' : 'static/[name].css',
-    chunkFilename: useHash ? '[name].[contenthash:8].chunk.css' : 'static/[name].chunk.css'
-  }])
+  chain.plugin('minify-css').use(MiniCssExtractPlugin, getBuildConfig().cssBuildConfig)
 
   chain.plugin('webpackBar').use(new WebpackBar({
     name: isServer ? 'server' : 'client',
@@ -235,9 +231,6 @@ const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
         resolvers: [ElementPlusResolver({
           ssr: isServer
         })]
-      }))
-      chain.plugin('ele3').use(require('unplugin-element-plus/webpack')({
-        lib: isServer
       }))
     }
   }