+++ packages/plugin-vue3/src/types.ts
@@ -1,9 +1,8 @@
-import type { VNode } from 'vue'
+import type { VNode, RendererNode, RendererElement } from 'vue'
 import type { Store, StoreOptions } from 'vuex'
 import type { RouteLocationNormalizedLoaded } from 'vue-router'
 import type { ISSRMidwayKoaContext, ISSRNestContext, ISSRContext, ESMFeRouteItem } from 'ssr-types'
 import type { Pinia } from 'pinia'
-import type { Readable } from 'stream'
 
 export interface ParamsKoa<T={}, U={}> {
   store: Store<T>
@@ -47,9 +46,17 @@ export interface VueRouterOptions {
   base?: string
 }
 
-export type Vue3RenderRes = Readable | {
-  html: string
-  teleportsContext: {
-    teleports?: Record<string, string> | undefined
-  }
+export interface vue3AppParams {
+  combineAysncData: any
+  state: any
+  layoutFetchData: any
+  asyncData: any
+  manifest: Record<string, string | undefined>
+  isCsr: boolean
+  jsInject: Array<VNode<RendererNode, RendererElement, {
+    [key: string]: any
+  }>>
+  cssInject: Array<VNode<RendererNode, RendererElement, {
+    [key: string]: any
+  }>>
 }