+++ packages/utils/src/server/static.ts
@@ -7,7 +7,7 @@ window.__vite_plugin_react_preamble_installed__ = true`
 const remInitial = "var w = document.documentElement.clientWidth / 3.75;document.getElementsByTagName('html')[0].style['font-size'] = w + 'px'"
 // module need be external
 const defaultExternal = ['@vue/server-renderer', 'ssr-serialize-javascript',
-  'react', 'react-dom', 'react-dom/server', 'vue', 'vue-router', 'vuex', 'pinia',
+  'react', 'react-dom', 'react-dom/server', 'vue', 'vue-router', 'vuex', 'pinia', 'ssr-react-dom',
   'ssr-server-utils', 'ssr-deepclone', 'ssr-hoc-react', 'ssr-common-utils', 'vite', 'axios']
 
 const ssrPackages = [
@@ -15,7 +15,7 @@ const ssrPackages = [
   'ssr-plugin-vue', 'ssr-types', 'ssr-common-utils',
   'ssr-core-vue', 'ssr-hoc-vue3', 'ssr-webpack', 'ssr-core-react',
   'ssr-core-vue3', 'ssr-plugin-midway', 'ssr-plugin-nestjs',
-  'ssr-hoc-react'
+  'ssr-hoc-react', 'ssr-core'
 ]
 
 export {